+++ src/ir/process.rs
@@ -5,8 +5,8 @@
 use crate::{
     ir::{
         Block, BlockData, ControlFlowGraph, DataFlowGraph, ExtUnit, ExtUnitData, FunctionInsertPos,
-        FunctionLayout, Inst, InstData, InstLayout, Signature, Unit, UnitBuilder, UnitKind,
-        UnitName, Value, ValueData,
+        FunctionLayout, Inst, InstData, Signature, Unit, UnitBuilder, UnitKind, UnitName, Value,
+        ValueData,
     },
     ty::Type,
     verifier::Verifier,
@@ -142,14 +142,6 @@ impl Unit for Process {
 
     fn func_layout_mut(&mut self) -> &mut FunctionLayout {
         &mut self.layout
-    }
-
-    fn inst_layout(&self) -> &InstLayout {
-        panic!("inst_layout() called on process");
-    }
-
-    fn inst_layout_mut(&mut self) -> &mut InstLayout {
-        panic!("inst_layout_mut() called on process");
     }
 
     fn dump_fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {